requirements.txt
asttokens==2.4.1
av==12.3.0
Brotli==1.1.0
certifi==2024.7.4
charset-normalizer==3.3.2
coloredlogs==15.0.1
comm==0.2.2
ctranslate2==4.3.1
debugpy==1.8.5
decorator==4.4.2
executing==2.0.1
ez_setup==0.9
faster-whisper==1.0.3
ffmpeg-python==0.2.0
filelock==3.15.4
flatbuffers==24.3.25
fsspec==2024.6.1
future==1.0.0
huggingface-hub==0.24.5
humanfriendly==10.0
idna==3.7
imageio==2.35.0
imageio-ffmpeg==0.5.1
ipykernel==6.29.5
ipython==8.26.0
jedi==0.19.1
jupyter_client==8.6.2
jupyter_core==5.7.2
lazy_loader==0.4
matplotlib-inline==0.1.7
moviepy==1.0.3
mpmath==1.3.0
mutagen==1.47.0
nest-asyncio==1.6.0
networkx==3.3
numpy==1.26.4
onnxruntime==1.19.0
opencv-python==4.10.0.84
packaging==24.1
parso==0.8.4
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.2.2
proglog==0.1.10
prompt_toolkit==3.0.47
protobuf==5.27.3
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pycryptodomex==3.20.0
pygame==2.6.0
Pygments==2.18.0
python-dateutil==2.9.0.post0
PyYAML==6.0.2
pyzmq==26.1.0
requests==2.32.3
scikit-image==0.24.0
scipy==1.14.0
setuptools==72.2.0
six==1.16.0
stack-data==0.6.3
sympy==1.13.2
tifffile==2024.8.10
tokenizers==0.20.0
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
typing_extensions==4.12.2
urllib3==2.2.2
wcwidth==0.2.13
websockets==13.0
yt-dlp==2024.8.6
